+++ apps/dht/h_find_successor.c
@@ -28,7 +28,7 @@ int h_find_successor(WOOF* wf, unsigned long seq_no, void* ptr) {
     log_set_level(DHT_LOG_INFO);
     // log_set_level(DHT_LOG_DEBUG);
     log_set_output(stdout);
-    srand(get_milliseconds());
+    srand((unsigned int)get_milliseconds());
 
     DHT_NODE_INFO node = {0};
     if (get_latest_node_info(&node) < 0) {
